LayoutTests/sputnik/Conformance/12_Statement/12.7_The_continue_Statement/S12.7_A8_T2-expected.txt

CONSOLE MESSAGE: line 87: SyntaxError: 'continue' is only valid inside a loop statement
S12.7_A8_T2

PASS Expected parsing failure

TEST COMPLETE
